requirements.txt
streamlit
databricks-sdk
openai>=1.0.0
